[joss-paper] Joan comments
Congrats on the paper, it reads really, really nice! I just emphasized the fact that EEG connectivity is done at the source level (vs. electrode) and I would suggest reviewing the second sentence of the first paragraph (It has revolutionized [...]), as it is a little bit cumbersome. Otherwise, it looks fantastic!